add HandleClickProps  in appTypes
